+++ cms/api/src/main/java/org/hippoecm/frontend/plugins/standards/list/DocumentsProvider.java
@@ -64,16 +64,16 @@ public class DocumentsProvider extends SortableDataProvider<Node> implements IOb
         this.observers = new HashMap<IModel<Node>, IObserver>();
     }
 
-    public Iterator<Node> iterator(int first, int count) {
+    public Iterator<Node> iterator(long first, long count) {
         load();
-        return entries.subList(first, first + count).iterator();
+        return entries.subList((int) first, (int) (first + count)).iterator();
     }
 
     public IModel<Node> model(Node object) {
         return new JcrNodeModel(object);
     }
 
-    public int size() {
+    public long size() {
         load();
         return entries.size();
     }
